+++ controllers/ReciboController.ts
@@ -1,21 +1,21 @@
 import { Request, Response } from "express";
 import mongoose from "mongoose";
-import { emitirReciboDeclaracao, lerConteudoPDF } from "../service/recibo/ReciboService";
+import { emitirReciboDeclaracao, lerConteudoPDF } from "../service/ReciboService";
 
 class ReciboController {
   async gerarRecibo(req: Request, res: Response): Promise<void> {
     try {
       const declaracaoId = mongoose.Types.ObjectId.createFromHexString(req.params.id);
-     
+
       const caminhoDeclaracao = await emitirReciboDeclaracao(declaracaoId);
 
-      
+
       const pdfConteudo = await lerConteudoPDF(caminhoDeclaracao);
 
       res.setHeader('Content-Type', 'application/pdf');
       res.setHeader('Content-Disposition', `attachment; filename="recibo.pdf"`);
 
-    
+
       res.status(200).send(pdfConteudo);
     } catch (error) {
       console.error("Erro ao gerar recibo:", error);
