+++ queue/ConsumerService.ts
@@ -1,15 +1,14 @@
 // No arquivo do consumidor da fila (o código que consome a fila)
 
 import amqp from "amqplib/callback_api";
-import xlsx from "xlsx";
-import Bibliografico from "../../models/Bibliografico";
-import Museologico from "../../models/Museologico";
-import Arquivistico from "../../models/Arquivistico";
-import Declaracoes from "../../models/Declaracao";
-import path from "path";
-import connectDB from "../../db/conn";
 import dotenv from "dotenv";
-import DeclaracaoService from "./declaracao/DeclaracaoService"; // Importar o serviço de declaração
+import path from "path";
+import xlsx from "xlsx";
+import connectDB from "../db/conn";
+import Arquivistico from "../models/Arquivistico";
+import Bibliografico from "../models/Bibliografico";
+import Museologico from "../models/Museologico";
+import DeclaracaoService from "../service/DeclaracaoService"; // Importar o serviço de declaração
 
 dotenv.config();
